added tests and restructured
