+++ lib/greetings.js
@@ -1,14 +1,3 @@
-var Botkit = require('botkit');
-
-var controller = Botkit.slackbot({
-  debug: false
-});
-
-var bot = controller.spawn({
-  token: process.env.PROJECT_GREETER_SLACK_BOT_TOKEN,
-  retry: 20
-}).startRTM()
-
 var ls_greet = "Welcome to LocalSupport <@USER_NAME> :slightly_smiling_face: Get started by <https://github.com/AgileVentures/LocalSupport/blob/develop/docs/installation.md#installation|installing the codebase> and our <#C3QFDHMAM|localsupport-install> channel if you need help"
 
 ls_greet += "\n\n\n see https://github.com/AgileVentures/LocalSupport/blob/develop/CONTRIBUTING.md for how to contribute to the project, but feel free to ask questions here too, flagging me with <@U0292M7AC|tansaku> :-)"
@@ -32,6 +21,7 @@ wed_greet += "\n\nSetup: https://github.com/WikiEducationFoundation/WikiEduDashb
 
 wed_greet += "\n\nIf you run into any trouble following the instructions for setting up a dev environment, please say something!"
 
+
 //for testing on staging slack
 // const greetings = { 
 //                     'C69J9GC1Y': ls_greet, 
@@ -48,23 +38,5 @@ const greetings = {
                     'C36MNPWTD': wed_greet
                   };
 
-controller.on('member_joined_channel', function(bot, message){
-  // console.log('member_joined_channel')
-  // bot.say({channel: message.channel, text: greet});
-  var greet = greetings[message.channel];
-  if(greet){
-    greet = greet.replace('USER_NAME', message.user);
-    bot.reply(message, greet);
-  }
-});
-
-
-// controller.on('message_received', function(bot, message) {
-//   console.log('message_received')
-//   console.log(message)
-// });
-
-
-
-
+module.exports = greetings;
